Add explicit data types

+++ grizli/model.py
@@ -342,7 +342,7 @@ class GrismDisperser(object):
         
         self.modelf = np.zeros(np.product(self.sh_beam), dtype=np.float)
         self.model = self.modelf.reshape(self.sh_beam)
-        self.idx = np.arange(self.modelf.size).reshape(self.sh_beam)
+        self.idx = np.arange(self.modelf.size, dtype=np.int64).reshape(self.sh_beam)
         
         ## Indices of the trace in the flattened array 
         self.x0 = np.array(self.sh, dtype=np.int64) // 2
